pyproject.toml
[project]
name = "home-exporter"
version = "0.1.0"
description = "🏠 Home InfluxDB Exporter"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "evohomeclient>=0.3.7",
    "influxdb3-python>=0.9.0",
    "lowatt-grdf>=2.3.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "sentry-sdk>=2.16.0",
]

[project.scripts]
home-exporter = "home_exporter:main"
# lint = "ruff check ."
# format = "ruff check . --fix"

[tool.uv]
package = true
dev-dependencies = [
    "ruff>=0.6.9",
]
